apas_verus::vstdplus::checked_int::checked_int

Trait CheckedIntTrait

pub trait CheckedIntTrait:
    View<V = int>
    + Sized
    + Clone {

    // Required methods
    spec fn spec_min() -> int;
    spec fn spec_max() -> int;
    exec fn is_normal(&self) -> normal : bool;
    exec fn is_out_of_range(&self) -> out_of_range : bool;
    exec fn add_checked(&self, other: &Self) -> sum : Self;
    exec fn sub_checked(&self, other: &Self) -> diff : Self;
    exec fn mul_checked(&self, other: &Self) -> product : Self;
    proof fn lemma_add_commutative_ghost(a: Self, b: Self);
    proof fn lemma_mul_commutative_ghost(a: Self, b: Self);
    proof fn lemma_sub_anticommutative_ghost(a: Self, b: Self);
    proof fn lemma_add_associative_ghost(a: Self, b: Self, c: Self);
    proof fn lemma_mul_distributes_over_add_ghost(a: Self, b: Self, c: Self);
    proof fn lemma_mul_distributes_over_sub_ghost(a: Self, b: Self, c: Self);

    // Provided method
    proof fn lemma_mul_associative_ghost(a: Self, b: Self, c: Self) { ... }

}

Trait for checked signed integer types that track overflow/underflow.

Weaker guarantees: both overflow AND underflow possible.

Cannot guarantee intermediate values stay in range even if final result does.

Required Methods

spec fn spec_min() -> int

spec fn spec_max() -> int

exec fn is_normal(&self) -> normal : bool

ensures
normal == (Self::spec_min() <= self@ <= Self::spec_max()),

exec fn is_out_of_range(&self) -> out_of_range : bool

ensures
out_of_range == !(Self::spec_min() <= self@ <= Self::spec_max()),

exec fn add_checked(&self, other: &Self) -> sum : Self

ensures
sum@ == self@ + other@,

exec fn sub_checked(&self, other: &Self) -> diff : Self

ensures
diff@ == self@ - other@,

exec fn mul_checked(&self, other: &Self) -> product : Self

ensures
product@ == self@ * other@,

proof fn lemma_add_commutative_ghost(a: Self, b: Self)

ensures
a@ + b@ == b@ + a@,

proof fn lemma_mul_commutative_ghost(a: Self, b: Self)

ensures
a@ * b@ == b@ * a@,

proof fn lemma_sub_anticommutative_ghost(a: Self, b: Self)

ensures
a@ - b@ == -(b@ - a@),

proof fn lemma_add_associative_ghost(a: Self, b: Self, c: Self)

ensures
(a@ + b@) + c@ == a@ + (b@ + c@),

proof fn lemma_mul_distributes_over_add_ghost(a: Self, b: Self, c: Self)

ensures
a@ * (b@ + c@) == a@ * b@ + a@ * c@,

proof fn lemma_mul_distributes_over_sub_ghost(a: Self, b: Self, c: Self)

ensures
a@ * (b@ - c@) == a@ * b@ - a@ * c@,

Provided Methods

proof fn lemma_mul_associative_ghost(a: Self, b: Self, c: Self)

ensures
(a@ * b@) * c@ == a@ * (b@ * c@),

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.